# Report ms/src/main/java/com/fms/api/ReportApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.16).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.fms.api;

import com.fms.model.Report;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-01-28T10:54:49.918+05:30[Asia/Calcutta]")
@Api(value = "report", description = "the report API")
public interface ReportApi {

    Logger log = LoggerFactory.getLogger(ReportApi.class);

    default Optional<ObjectMapper> getObjectMapper(){
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest(){
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "Get all report details", nickname = "allReport", notes = "All the reports are displayed", response = Report.class, responseContainer = "List", tags={ "Report Details", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful Operation", response = Report.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Invalid status value"),
        @ApiResponse(code = 404, message = "Page not found") })
    @RequestMapping(value = "/report",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Report>> allReport() {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("[ {\r\n  \"Status\" : \"Status\",\r\n  \"VenueAddress\" : \"VenueAddress\",\r\n  \"EventDate\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"TotalVolunteers\" : 0,\r\n  \"BeneficiaryName\" : \"BeneficiaryName\",\r\n  \"TotalTravelHours\" : 1,\r\n  \"CouncilName\" : \"CouncilName\",\r\n  \"EventName\" : \"EventName\",\r\n  \"Month\" : \"January\",\r\n  \"BusinessUnit\" : \"BusinessUnit\",\r\n  \"EventId\" : \"EventId\",\r\n  \"TotalVolunteerHours\" : 6,\r\n  \"BaseLocation\" : \"United Kingdom\"\r\n}, {\r\n  \"Status\" : \"Status\",\r\n  \"VenueAddress\" : \"VenueAddress\",\r\n  \"EventDate\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"TotalVolunteers\" : 0,\r\n  \"BeneficiaryName\" : \"BeneficiaryName\",\r\n  \"TotalTravelHours\" : 1,\r\n  \"CouncilName\" : \"CouncilName\",\r\n  \"EventName\" : \"EventName\",\r\n  \"Month\" : \"January\",\r\n  \"BusinessUnit\" : \"BusinessUnit\",\r\n  \"EventId\" : \"EventId\",\r\n  \"TotalVolunteerHours\" : 6,\r\n  \"BaseLocation\" : \"United Kingdom\"\r\n} ]", List.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default ReportApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "filter event", nickname = "filterEvent", notes = "", response = Report.class, tags={ "Report Details", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful Operation", response = Report.class),
        @ApiResponse(code = 400, message = "Invalid status value"),
        @ApiResponse(code = 404, message = "Page not found") })
    @RequestMapping(value = "/report",
        produces = { "application/xml", "application/json" }, 
        consumes = { "application/json", "application/xml" },
        method = RequestMethod.POST)
    default ResponseEntity<Report> filterEvent(@ApiParam(value = "Report that appears after filter" ,required=true )  @Valid @RequestBody Report body
) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{\r\n  \"Status\" : \"Status\",\r\n  \"VenueAddress\" : \"VenueAddress\",\r\n  \"EventDate\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"TotalVolunteers\" : 0,\r\n  \"BeneficiaryName\" : \"BeneficiaryName\",\r\n  \"TotalTravelHours\" : 1,\r\n  \"CouncilName\" : \"CouncilName\",\r\n  \"EventName\" : \"EventName\",\r\n  \"Month\" : \"January\",\r\n  \"BusinessUnit\" : \"BusinessUnit\",\r\n  \"EventId\" : \"EventId\",\r\n  \"TotalVolunteerHours\" : 6,\r\n  \"BaseLocation\" : \"United Kingdom\"\r\n}", Report.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default ReportApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}